Rename constants module
