+++ src/features/applicationView/ApplicationView.tsx
@@ -11,7 +11,7 @@ import CardHeader from '../../common/cardHeader/CardHeader';
 import Text from '../../common/text/Text';
 import { formatDate } from '../../common/utils/format';
 import Chip from '../../common/chip/Chip';
-import { APPLICATION_STATUS } from '../../common/utils/consonants';
+import { APPLICATION_STATUS } from '../../common/utils/constants';
 import CustomerProfileCard, { CustomerProfileCardProps } from '../../common/customerProfileCard/CustomerProfileCard';
 import BerthOfferCard, { BerthOfferCardProps } from './berthOfferCard/BerthOfferCard';
 import PageTitle from '../../common/pageTitle/PageTitle';
