+++ src/features/offer/tableTools/TableTools.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next';
 import styles from './tableTools.module.scss';
 import Text from '../../../common/text/Text';
 import Chip from '../../../common/chip/Chip';
-import { APPLICATION_STATUS } from '../../../common/utils/consonants';
+import { APPLICATION_STATUS } from '../../../common/utils/constants';
 import { ApplicationStatus } from '../../../@types/__generated__/globalTypes';
 import Button from '../../../common/button/Button';
