+++ src/features/unmarkedWsNoticeList/UnmarkedWsNoticeList.tsx
@@ -7,7 +7,7 @@ import InternalLink from '../../common/internalLink/InternalLink';
 import Table, { Column, COLUMN_WIDTH } from '../../common/table/Table';
 import { formatDate } from '../../common/utils/format';
 import Chip from '../../common/chip/Chip';
-import { APPLICATION_STATUS } from '../../common/utils/consonants';
+import { APPLICATION_STATUS } from '../../common/utils/constants';
 import { ApplicationStatus } from '../../@types/__generated__/globalTypes';
 import { UnmarkedWinterStorageNotice } from './utils';
 import { SortedCol } from '../../common/utils/useBackendSorting';
