+++ src/features/winterStorageApplicationList/WinterStorageApplicationList.tsx
@@ -8,7 +8,7 @@ import InternalLink from '../../common/internalLink/InternalLink';
 import Table, { Column, COLUMN_WIDTH } from '../../common/table/Table';
 import { formatDate } from '../../common/utils/format';
 import Chip from '../../common/chip/Chip';
-import { APPLICATION_STATUS } from '../../common/utils/consonants';
+import { APPLICATION_STATUS } from '../../common/utils/constants';
 import { ApplicationStatus } from '../../@types/__generated__/globalTypes';
 import { WinterStorageApplication } from './utils';
 import ApplicationDetails from '../../common/applicationDetails/ApplicationDetails';
